refactor: move discounts plugin to Node app
