+++ imports/node-app/core-services/discounts/i18n/fr.json
