+++ imports/node-app/registerPlugins.js
@@ -1,6 +1,7 @@
 /* node-app imports */
 /* core-services */
 import registerCatalogPlugin from "./core-services/catalog/index.js";
+import registerDiscountsPlugin from "./core-services/discounts/index.js";
 import registerEmailPlugin from "./core-services/email/index.js";
 import registerFilesPlugin from "./core-services/files/index.js";
 import registerI18nPlugin from "./core-services/i18n/index.js";
@@ -42,7 +43,6 @@ import registerCartPlugin from "/imports/plugins/core/cart/server/no-meteor/regi
 import registerCheckoutPlugin from "/imports/plugins/core/checkout/server/no-meteor/register";
 import registerCorePlugin from "/imports/plugins/core/core/server/no-meteor/register";
 import registerDiscountCodesPlugin from "/imports/plugins/included/discount-codes/server/no-meteor/register";
-import registerDiscountsPlugin from "/imports/plugins/core/discounts/server/no-meteor/register";
 import registerJobQueuePlugin from "/imports/plugins/included/job-queue/server/no-meteor/register";
 import registerMarketplacePlugin from "/imports/plugins/included/marketplace/server/no-meteor/register";
 import registerOrdersPlugin from "/imports/plugins/core/orders/server/no-meteor/register";
