--- imports/plugins/core/discounts/server/api/index.js
@@ -1,3 +1,0 @@
-import Reaction from "./import";
-
-export default Reaction;
